+++ lib/src/models/model_exports.dart
@@ -7,3 +7,5 @@ export 'package:universal_ble/src/models/ble_service.dart';
 export 'package:universal_ble/src/models/availability_state.dart';
 export 'package:universal_ble/src/models/ble_connection_state.dart';
 export 'package:universal_ble/src/models/ble_device.dart';
+export 'package:universal_ble/src/models/ble_command.dart';
+